Mas Longs cambiados por int, modificacion en BotHelper cambiar chatId a int para funcionamiento

+++ oci-react-samples/MtdrSpring/backend/src/main/java/com/springboot/MyTodoList/service/AuthenticationService.java
@@ -13,7 +13,7 @@ public class AuthenticationService {
     @Autowired
     private OracleUserRepository oracleUserRepository;
 
-    public OracleUser authenticate(Long userId) {
+    public OracleUser authenticate(int userId) {
         Optional<OracleUser> user = oracleUserRepository.findById(userId);
         return user.orElse(null);
     }
